abstracting out the navlink variables
